alterações nas classes básicas
add construtores, setters de validação, alguns equals...
