--- Livraria_Rural/src/Administrador.java
@@ -1,3 +1,0 @@
-public class Administrador extends Pessoa{
-
-}
