--- Livraria_Rural/src/Pessoa.java
@@ -1,41 +1,0 @@
-public class Pessoa {
-    protected String nome;
-    protected String email;
-    protected String senha;
-    protected String telefone;
-
-    // Getters e Setters
-
-
-    public String getNome() {
-        return nome;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getSenha() {
-        return senha;
-    }
-
-    public String getTelefone() {
-        return telefone;
-    }
-
-    public void setTelefone(String telefone) {
-        this.telefone = telefone;
-    }
-
-    public void setSenha(String senha) {
-        this.senha = senha;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-}
