--- Livraria_Rural/src/RepositorioLivro.java
@@ -1,2 +1,0 @@
-public class RepositorioLivro {
-}
