--- Livraria_Rural/src/RepositorioPessoa.java
@@ -1,2 +1,0 @@
-public class RepositorioPessoa {
-}
