--- Livraria_Rural/src/RepositorioVenda.java
@@ -1,2 +1,0 @@
-public class RepositorioVenda {
-}
