+++ Livraria_Rural/src/beans/Administrador.java
@@ -1,0 +1,25 @@
+package beans;
+
+import java.time.LocalDate;
+
+public class Administrador extends Pessoa {
+
+    // Atributos
+
+    // Construtor
+    public Administrador(String nome, String email, String senha,
+                         String telefone, LocalDate dataNascimento) {
+        super(nome, email, senha, telefone, dataNascimento);
+    }
+
+    // Metodos
+
+    // equals
+
+    // toString
+
+    // Getters
+
+    // Setters
+
+}
