+++ Livraria_Rural/src/beans/Genero.java
@@ -1,0 +1,7 @@
+package beans;
+
+public enum Genero {
+
+    GENERO1, GENERO2, GENERO3, GENERO4
+
+}
