+++ Livraria_Rural/src/dados/RepositorioLivro.java
@@ -1,0 +1,4 @@
+package dados;
+
+public class RepositorioLivro {
+}
