+++ Livraria_Rural/src/dados/RepositorioPessoa.java
@@ -1,0 +1,4 @@
+package dados;
+
+public class RepositorioPessoa {
+}
