+++ Livraria_Rural/src/dados/RepositorioVenda.java
@@ -1,0 +1,4 @@
+package dados;
+
+public class RepositorioVenda {
+}
